workaround for issue #54
This is a temporary workaround for https://github.com/UltraCart/responsive_checkout/issues/54. It’s not great but it gets the job done. It adds the model’s cid as the id of each item; when the remove button is activated, this sets the quantity of that item to 0, initiates a cart save, hides this item, and once the cart save completes, reloads the page to prevent an extra incorrect item from showing up.